[mono-runtimes, monodroid] Fix Mono runtime building, dependencies (#83)
Commit 6bb1016 "broke" `build-tools/mono-runtimes`, in that it didn't
build any Android target JIT ABIs, because it used the wrong MSBuild
property for the condition checks -- it used
`$(AndroidSupportedTargetJitAbis)`, not
`$(AndroidSupportedTargetJitAbisForConditionalChecks)`. Doh!

Fix that.

Additionally, `src/monodroid` needs to depend upon
`build-tools/mono-runtimes` because it depends upon eglib-config.h,
which is generated as part of the mono build. `src/monodroid` doens't
depend upon the generated `libmonosgen-2.0.so` file, but it *does*
depend upon the build tree.