[generator-Tests] Provide more context when errors happen (#48)
A unit test failure is being reported, which unfortunately doesn't
contain enough useful context:

	1) generatortests.NormalMethods.GeneratedOK :   error : .../Java.Interop/bin/TestDebug/SupportFiles/JavaObject.cs(78,17): (Location of the symbol related to previous warning)
	Expected: 0
	But was:  1

It's a very odd "error", in that there's no error reported, just a
mention of a "previous warning", which isn't shown.

The *cause* of this useless output is that `Compiler` "trims out" all
the "non-error" output from the compilation. This *seems* good, except
in this case it's the "useless output" that we need to make sense of
anything at all!

Revise `Compiler.Compile()` so that instead of extracting "errors" it
instead provides *all* compiler output and an indicator of whether or
not errors occurred. (This still depends on `CompilerResults.Errors`
having correct output -- and I'm starting to wonder if I'm actually
dealing with a CodeDom bug here -- but at least with more output I can
better rationalize about what's going on.)