Automatically install build dependencies on Linux
The only distribution we haved support for right now is Ubuntu but it's
trivial to add support for your distro of choice. To do so, run on
command line:

  lsb_release -i -s

and take note of the printed string - this is your distribution name as
known to LSB. If the installation commands work for all releases of your
distribution, add the following target to the Makefile:

  linux-prepare-YOUR_DISTRO_NAME::

and provide commands to install the dependencies using your package
manager below that line.

If, however, some releases of your distribution require different
commands or a different set of packages to install, run on command line:

  lsb_release -r -s

And take note of the printed distribution release string. Then add the
following target to the Makefile:

  linux-prepare-YOUR_DISTRO_NAME-DISTRO_RELEASE::

And omit the first target above. That's it!