[Xamarin.Android.Build.Tasks] Use libZipSharp to build the .apk (#89)
Commit 4ec06ac9 *broke* generation of `.apk` files when migrating
from `Ionic.Zip.dll` to `System.IO.Compression` because this:

	apk.AddFile (assembly.ItemSpec, "assemblies", compressionLevel: CompressionLevel.NoCompression);

doesn't work the way we thought/hoped it would.

Specifically, we require that assemblies be *stored*, uncompressed,
within the `.apk`, as we **mmap**(2) the `.apk` and hand off the
memory addresses of the loaded assemblies to mono for execution.

`CompressionLevel.NoCompression`, despite saying "no compression",
does *not* mean "store". It means "add the file 'normally', but tell
`DeflateStream` to use 'no compression'."

The result is that the zip entry doesn't say it's stored:

	$ unzip -lv bin/Debug/*-Signed.apk | grep assemblies/
		 64512  Defl:N    64512  67%  06-16-16 10:38  6d65706b  assemblies/Scratch.DebugRelease.dll

Which in turn means that the resulting `.apk` is *unusable*.

The fix? `System.IO.Compression` can't be used for this, and
`Ionic.Zip.dll` had other problems (why we tried to use
`System.IO.Compression` in the first place!), so instead we'll
use [libZip][0] and [LibZipSharp][1] to handle `.apk` files.
libzip is a well-maintained OSS library for manipulating zip files,
and LibZipSharp is a C# wrapper around libzip.

Add libzip and LibZipSharp to the build system so that they're
available on required paltforms, and update the `BuildApk` and
related tasks to use LibZipSharp instead of System.IO.Compression.
This allows assemblies to be properly stored in the `.apk`,
allowing apps to execute as intended.

	$ unzip -lv bin/Debug/*-Signed.apk | grep assembl
		 64512  Stored    64512   0%  06-29-16 08:31  6d65706b  assemblies/Scratch.DebugRelease.dll

[0]: http://www.nih.at/libzip/
[1]: https://github.com/grendello/LibZipSharp/