Fix an issue where all the assemblies are added with the name assembies.
In the commit f3d62b6  which switches over to using libzip there was
an oversight. The new API needs the filename in the archive
not just the directory (like IonicZip did). As a result the code
was making ALL the assemblies have the same name.

This commit fixes that issue by appending the filename to the
"assemblies" path.

It also fixes an issue with the typemap files not being added
to the root of the apk.