[Mono.Android] remove IAudioRouting as the supported API binding so far.
It caused API breakage because of the "newly added" base interface for
routing change listener, and its relevant event handler generation totally
made it incompatible with the former API. Namely, it killed
RoutingChangedEventArgs type and relevant events.

Remove AudioRouting and relevant types/members from the binding to preserve
API compatibility.

I don't think it is possible forever to bind this AudioRouting type and
it is going to be problematic in the future as long as people weigh on
API compatibility mantra.