[Mono.Android] bring back some Java.Nio channel classes that were removed.
N API brought in a new interface java.nio.channels.NetworkChannel and
DatagramChannel, SocketChannel and ServerSocketChannel now implement it.

However this interface cannot be bound due to generics limitation, so
we disable it.

When we disabled it, we thought that it there would be only new types
that would implement this new interface. But it is actually implemented
by *existing* classes as mentioned above.

So, instead of eliminating the entire tree, eliminate <implements> elements
on those types instead.