[monodroid] Fix COMMON_CFLAGS usage. (#82)
* Fix monodroid COMMON_CFLAGS usage and always build with RELEASE.

The primary problem was that when app launches it tries to load assemblies:

	W/monodroid( 2061): Using override path: /data/data/com.xamarin.android.helloworld/files/.__override__

It should not happen, because we don't have fastdev here.

However, it was weird, that the problem occured even with app Release builds.

Turned out that in src/monodroid, Android.mk referenced COMMON_FLAGS,
which did not exist. Fix typo.

Then, since we don't have fastdev, monodroid needs to be built always
in RELEASE mode. So, add -DRELEASE=1 even for Debug build.

* Remove RELEASE build flag from common. Instead default xabuild to Release.