Applications crashing on startup due to java.io.IOException: No original dex files found (#115)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=42521

Recent changes to the AddFiles method in LibZipSharp ment that
the fix to strip the path of the classes.dex file was reverted.
This commit fixes that. We bump LibZipSharp to bring in a fix
where if we do not supply a path and useFilenamesPath is false
we only use the Filename for the zip.