[Xamarin.Android.Build.Tasks] Fix a regression in the Path seperator unit tests
No idea why this started happening. Its probably a change in
behaviour under Mono 4.4/.x version of xbuild. since it only
effects items using the "LigicalName" meta data.
But the result is we get things like this in the Resource.Designer.cs

	static int drawable/foo = 123456789;

which is obviously a compilation problem for c#.
So to fix we add some processing of the logical Name to replace
any "invalid" path seperators with valid ones.