[Xamarin.Android.Build.Tasks] .targets file should be an Inputs to cache-generating tasks
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=42671

When Xamarin.Android is updted we really should invalidate the
sdk cache just in case things changes/updated.

This commit adds the `Xamarin.Android.Common.targets` to the
list of inputs for the `_BuildSdkCache` target. This means if
the .targets file is newer than the Sdk cache it will auto
rebuild it.