Fix periodic crash on Monodroid appdomain unload
When the Designer unloads the Monodroid assembly, mono_java_gc_bridge_info continues pointing to data in the now-invalid assembly image. This currently causes a semi-consistent crash in the Designer tests because the appdomain is unloaded and then reloaded, create_domain (which creates the AppDomain managed object, and therefore implicitly depends on mono_java_gc_bridge_info) is called before init_android_runtime (which populates mono_java_gc_bridge_info with the new data).

The solution is to clear mono_java_gc_bridge_info when Monodroid is unloaded. This means all objects will be marked transparent until a new Monodroid is loaded, which is fine because only pure C# objects are being created anyway if Monodroid is unloaded (with one edge case which may still need to be fixed-- see the FIXME comment in this commit).