[build] Fix `make prepare`
Commit 29addc7f broke `make jenkins`:

	build-tools/scripts/Paths.targets: error :
	.../xamarin-android/build-tools/scripts/Paths.targets:
	Exception has been thrown by the target of an invocation.

The cause was due to the use of `build-tools/scripts/Paths.targets`
from `make prepare-external`, which runs before `make prepare-props`,
and thus before `Configuration.OperatingSystem.props` is created.

Not including `Configuration.OperatingSystem.props` in turn meant that
`$(HostOS)` wasn't set, and thus `$(AndroidMxeInstallPrefix)` wasn't
set, causing the `Path.GetFullPath()` for `$(AndroidMxeFullPath)` to
throw, and nothing worked.

Fix that by altering `$(AndroidMxeFullPath)` to be conditional on
`$(AndroidMxeInstallPrefix)` containing a value, and make importing of
`Configuration.OperatingSystem.props` conditional on the
`$(DoNotLoadOSProperties)` MSBuild property *not* being True.
(By default, it won't be set, and thus
`Configuration.OperatingSystem.props` will be included.)