[msbuild]: Add TLS Provider selection for BTLS.
* Tasks/BuildApk.cs: Add 'TlsProvider' parameter.

* Xamarin.Android.Common.targets: Add <AndroidTlsProvider> property
  and pass it to BuildApk.

Add this to your .csproj file to select the TLS Provider.  Valid
values are "legacy" and "btls".

We now add a new 'XA_TLS_PROVIDER' environment variable to the .apk,
which is read by System/Mono.Net.Security/MonoTlsProviderFactory.Droid.cs
to select the TLS Provider.

Linker support (to eliminate the unused provider, similar to how it's
already done in xamarin-macios) is coming shortly.