Bump to LibZipSharp/91e73c17 (#134)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=42835

Android projects which use the Support packages are failing to
build on Windows due an issue where IsDirectory was returning
false for directories.

This is because while the Size is 0 the compressed size
is NOT 0. This is probably because the data has to be stored
somehow in the zip :).

So for directories we'll check that the Size is zero and that
the string ends in a '/'

Ths commit Bumps the submodule for LibZipSharp to bring in the fix.