[Xamarin.Android.Build.Utilities] Xamarin can't find NDK 12b because ndk-stack.exe is now ndk-stack.cmd
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=42566

In the latest NDK release google decided to change ndk-stack.exe to
nkd-stack.cmd. As a result our ndk discovery code failed on Windows.

This commit updates the code to look for all files with contain
one of the %PATHEXT% extentions.

	.COM;.EXE;.BAT;.CMD;.VBS;.VBE;.JS;.JSE;.WSF;.WSH;.MSC

so that we are future proofed in case google decide to use .bat
next time.