[libmonodroid] Fix context switching with multithread access.
Previously, we were using the value of `mono_domain_get` and comparing it to our context
corresponding appdomain instance to decide if we had to reset the Java TypeManager integration
(called via the `reinitialize_android_runtime_type_manager` function).

When instantiating several sessions and refreshing them rapidly at the same time (which is
something the Xamarin.Forms previewer does), we started noticing cases of objects being created
in one domain and subsequently being used in another although the domain context should have
been switched.

I tracked down the issue to TypeManager.java and the `n_activate` method which was, in this
case, incorrectly allocating the managed peer in a different domain than it should.

The reason is because the designer hosts its different sessions in individual thread, contexts
are locally created and used from different execution contexts. As `mono_domain_get` uses
Mono's TLS to store which domain it's currently on, the function is *not* actually tracking a
global domain usage from native code but rather a per-thread usage.

This meant that TypeManager.java `registerJniNatives` method was silently not being called because
from the session thread point of view, it was already on the right domain and that in turn prevented
subsequent type creation from happening in their rightful domain because the glue code between Java
and managed hasn't been reset to tap into that domain.

Thus to achieve the desired effect, this commit introduces a new field to track which context
ID was last selected so that we can make sure the Java integration is properly reset anytime
it needs to be.