[Xamarin.Android.Build.Tasks] Update the Aot task to use the correct msym parameters. (#143)
The mono team changed the way you tell the aot compiler to generate
the .msym files [1]. This old parameter has been replaced wtih a new
`msym-dir` parameter.

Rather than specifying the actual filename we now have to
specify the directory to output the data. Also note that the
.msym files don't end up in the root of this directory. You end up
with a bunch of hashed subdirectories which contain a .msym file
for each of the assemblies that is AoT'd.
Also note the MONO_DEBUG setting has also been deprecated.

Once this commit has been merged the Unit tests will need to be
updated to take into account the locations of the new files. That
PR will also include the required bump for this commit.

[1] https://github.com/mono/mono/tree/master/mcs/tools/mono-symbolicate