[Xamarin.Android.Build.Tasks] Add Items to _SharedRuntimeAssemblies to ensure they get included in Profile.g.cs
There was a bug where items like OpenTK and Mono.Android were not
being included in the Profile.g.cs file. As a result they were
not being picked up as framework assemblies. This resulted in
duplicate java classes being generated at build time.

This commit fixes this by manually adding the required items
to the _SharedRuntimeAssemblies ItemGroup. This works because
we don't do any file exists checks on the files in this ItemGroup.