Make package installation on Linux optional
Xamarin.Android needs a number of packages to be installed on Linux
in order to build properly. The makefile provides support for that but
it requires using `sudo` to invoke the package installer. That may not
be desired on build bots and so this commit makes package installation
optional. In order to avoid package installation invoke make like so:

   make NO_SUDO=true prepare

In this mode the makefile will merely report the installation status of
all the dependency packages instead of installing them.