[build] Package build artifacts with `make jenkins`
Update the `make jenkins` target to package up the `bin` directory for
(hopeful) use on other machines.

The generated package is named:

	oss-xamarin-android_vVERSION_OS-ARCH_BRANCH_COMMIT.zip

where:

* **VERSION** is the new `$(ProductVersion)` MSBuild property
    along with the number of commits since `$(ProductVersion)`
    was *changed*
* **OS** is the name of the OS that built the package
* **ARCH** is the OS architecture
* **BRANCH** is the git branch being built (usually `master`)
* **COMMIT** is the git commit that was built.

For example:

	oss-xamarin.android_v7.0.99.0_Darwin-x86_64_jonp-zip-output_c4eab00.zip

The package file is *not* an installer, just a ZIP package, but should
contain everything needed to *create* an installer.

This also copies `tools/scripts/xabuild` into the appropriate
`bin/$(Configuration)/bin` directories so that an `xabuild` script is
installed as part of the package, and `xabuild` is updated to setup
the various environment variables accordingly so it is invokable from
the `bin/$(Configuration)/bin` directories.