Allow MonoGCBridgeSCC objects to have an empty User Peer list
Mono passes gc_cross_references a description of a graph. Each node in
the graph has a list of User Peer objects which should be collected
together in the upcoming GC. This list is required by both the spec
and the current code to be nonempty. To enable new optimizations, the
runtime team wants to modify the spec to allow the list to be empty.
This patch makes that possible by creating a new temporary object of
type GCUserPeer when a MonoGCBridgeSCC with an empty list is seen.