[build] Include zip basename directory in build artifacts
Commit 1eebbe33 introduces generation of `oss-xamarin-android*.zip`
files when using the `make jenkins` target.

There's one "problem" with this file: a common convention for .zip
files is that they will contain a single directory named as the
basename of the .zip file itself:

	$ unzip -l foo-1.zip
	... foo-1/
	... foo-1/file1
	... foo-1/file2
	...

The benefit to this scheme is that if you have multiple "versions" of
the file, you can extract them without fear of file conflicts:

	# extracts into foo-1
	$ unzip foo-1.zip

	# extracts into foo-2
	$ unzip foo-2.zip

The `oss-xamarin-android*.zip` packages *didn't* follow this scheme:
instead, they contained the `bin` directory as it's top directory:

	$ unzip -l oss-xamarin.android_v7.0.99.2_Darwin-x86_64_HEAD_fbfd676.zip  | head -10
	Archive:  /Users/jon/Downloads/oss-xamarin.android_v7.0.99.2_Darwin-x86_64_HEAD_fbfd676.zip
	  Length     Date   Time    Name
	 --------    ----   ----    ----
	        0  08-10-16 21:46   bin/
	        0  08-10-16 23:40   bin/BuildDebug/
	    43008  08-10-16 20:10   bin/BuildDebug/api-merge.exe

This means you can't download and extract both
`oss-xamarin.android_v7.0.99.1*.zip` and
`oss-xamarin.android_v7.0.99.2*.zip` into the same directory -- not
without extra work -- which is counter to convention.

Fix the `make package-oss` target so that the zip's basename is
included in the directory structure:

	$ unzip -l oss-xamarin.android_v7.0.99.2_Darwin-x86_64_HEAD_fbfd676.zip  | head -10
	Archive:  /Users/jon/Downloads/oss-xamarin.android_v7.0.99.2_Darwin-x86_64_HEAD_fbfd676.zip
	  Length     Date   Time    Name
	 --------    ----   ----    ----
	        0  08-10-16 21:46   oss-xamarin.android_v7.0.99.2_Darwin-x86_64_HEAD_fbfd676/bin/
	        0  08-10-16 23:40   oss-xamarin.android_v7.0.99.2_Darwin-x86_64_HEAD_fbfd676/bin/BuildDebug/
	    43008  08-10-16 20:10   oss-xamarin.android_v7.0.99.2_Darwin-x86_64_HEAD_fbfd676/bin/BuildDebug/api-merge.exe