[Xamarin.Android.Build.Tasks] Fixed F# Resource Designer compilation issues
The latest release of F# removed support for static fields. As a result
the Resource.Designer.fs file that wes being generated by our build
system no longer compiled.

This commit reworks the Designer.cs code to emit a C# assembly for
F# projects rather than just the code. This has the benifit of
us not having to produce F# code in the first place.

The new system works as follows.

1) GenerateResourceDesigner Task generates a Resources.Designer.cs file as normal
2) If the target project is F# that file is then compiled into a Resource.Designer.dll
3) The new .dll is then included in the @ReferencePaths ItemGroup
4) The Resource.Designer.cs is removed from the @Complie ItemGroup

If the project is C# or VS the Designer file is just included as normal.