[mono-runtimes] Use directory separators with `$(OutputPath)`.
[The build][0] [is broken][1]:

	xamarin-android/build-tools/mono-runtimes/mono-runtimes.targets:
	error : Could not find a part of the path
	".../xamarin-android/xamarin-android/bin/Debuglib/xbuild-frameworks/MonoAndroid/v1.0/RedistList/FrameworkList.xml".

The error is generated from the `_InstallBcl` target when writing
`FrameworkList.xml`, and a close reading of the error message shows
that a directory separator is missing: it should be `bin/Debug/lib`,
but is instead `bin/Debuglib`.

I *assume* that this is due to `<CallTarget/>` being "flakey" within
xbuild (it passes the file containing the target to `<MSBuild/>`!),
and *somehow* in that "flakiness" `$(OutputPath)` doesn't end with
`\`, *which it should*, because `Microsoft.Common.targets` will
automatically append `\` to `$(OutputPath)`.

To work around this issue, update `mono-runtimes.targets` so that it
*always* uses `\` after `$(OutputPath)` and
`$(IntermediateOutputPath)`.

Additionally, `llc.exe` was being installed into the wrong directory,
preventing the bundle from being used:

	Target _BuildUnlessCached needs to be built as output file '../../bin/Debug/lib/mandroid/llc.exe' does not exist.

`llc.exe` was instead installed into `bin/Debug/bin/llc.exe`.

Bump the bundle version number to -v3 and correct
`mono-runtimes.targets` so that `llc.exe` and `opt.exe` are properly
installed into `$prefix/lib/mandroid`.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/49/
[1]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/49/consoleText