[libzip] Fix `<CallTarget/>` use
[The build][0] [is broken][1], while building `libzip.mdproj`:

	xamarin-android/build-tools/scripts/RequiredPrograms.targets: error :
	Error executing task Which:
	Required property 'Program' not set.

...to which I reply... Wat? [2]

`@(RequiredPrograms)` is set in `libzip.projitems`, and none have an
empty value, so how is this possible?

For now, I'm going to assume it's a variation on the xbuild
`<CallTarget/>` bug mentioned in 57b18c27, and apply a similar fix:
make `libzip.targets` "stand alone" so that it imports everything it
requires (compared with the current situation, in which
`libzip.mdproj` does most of the importing).

This will hopefully appease the xbuild deities.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/50/
[1]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/50/consoleText
[2]: https://www.destroyallsoftware.com/talks/wat