Instead of a lookaside for temporary peer indices, memoize on SCCs
In the GC bridge handling, we need to be able to map bridgeless SCCs
to their index in the temporary_peers array. Currently we store the
indices in a lookaside table. This patch instead hides the index in
a field in MonoGCBridgeSCC which is unused for bridgeless SCCs,
allowing us to avoid allocating the lookaside.