F# projects don't build when lang is changed in XS
Setting the UI language in XS sets the msbuild variable LANGUAGE = en
(or whatever language you choose).

The Android F# targets check for the existence of this variable to
determine if Microsoft.FSharp.targets has been loaded.

Microsoft.FSharp.targets sets <Language>F#</Language> so it is better
to check for the presence of F# being set rather than checking to see
if it's not empty.

Fixes #43525