[Xamarin.Android.Build.Tasks] Extend the Aot Task to allow the passing of additional arguments (#185)
Context https://bugzilla.xamarin.com/show_bug.cgi?id=43500

We need a way to pass additional arguments to the aot system.
This is so users can add additional parameters to tweak the
output. In the case of Bug 43500 we needed to be able to pass

	'no-write-symbols' or 'no-debug'

Rather than just hardcoding those values we now provide a

	$(AndroidAotAdditionalArguments)

msbuild property which defaults to empty but users can override
to pass any additional parameters that they need.