[Xamarin.Android.Build.Utilities] Xamarin can't use NDK 12b as ndk-stack.exe is now ndk-stack.cmd (#188)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=42566

The problem was that we did not UPDATE the ndk-stack.exe property
to be ndk-stack.cmd once we had verified that the ndk was valid.
Seeing as this is not the first time google has changed an extension
on a executable we need to do some re-work on these classes.

Firstly we'll default all the values to the 'unix' ones. We will
then use the new Executables method to search for the other exe's
on initalization and update the properties accordinly.

However because some of the files (the java related ones) are in
other paths (like build-tools/x.x.x.x) we will override those
on windows to give them the exepcted .exe extension.