[Xamarin.Android.Build.Tasks] [Symbolication] Managed assemblies and mdbs should be stored in MVID folders
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=43550

The mono-symbolication now provides a way to copy the assembly files
into the correct directories. Before we were just dumping them in
the root of the msym archive. But the spec changed so that each
assembly needed to be in its own MVID folder.

Rather than trying to figure out that we will do what the iOS team
has done. We call mono-symbolicate with the `store-symbols` parameter
passing it the msym archive directory and the directory where all
the intermediate msym files are.

This makes sure that the files are formatted and placed in the
correct places.